requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pymongo[srv]==4.6.1
dnspython==2.4.2
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.1
cryptography==41.0.7
motor==3.3.2
urllib3==2.1.0